Move :: move exception package location to global
