+++ src/main/java/com/keepgoing/keepserver/domain/user/controller/UserController.java
@@ -1,7 +1,7 @@
 package com.keepgoing.keepserver.domain.user.controller;
 
 import com.keepgoing.keepserver.domain.user.entity.user.User;
-import com.keepgoing.keepserver.domain.user.exception.BusinessException;
+import com.keepgoing.keepserver.global.exception.BusinessException;
 import com.keepgoing.keepserver.domain.user.payload.request.LoginRequest;
 import com.keepgoing.keepserver.domain.user.payload.request.SignupRequest;
 import com.keepgoing.keepserver.domain.user.payload.request.UserInfoRequest;
