--- src/main/java/com/keepgoing/keepserver/domain/user/exception/error/ErrorProperty.java
@@ -1,8 +1,0 @@
-package com.keepgoing.keepserver.domain.user.exception.error;
-
-import org.springframework.http.HttpStatus;
-
-public interface ErrorProperty {
-    HttpStatus getStatus();
-    String getMessage();
-}
