+++ src/main/java/com/keepgoing/keepserver/domain/user/service/user/UserService.java
@@ -1,6 +1,6 @@
 package com.keepgoing.keepserver.domain.user.service.user;
 
-import com.keepgoing.keepserver.domain.user.exception.BusinessException;
+import com.keepgoing.keepserver.global.exception.BusinessException;
 import com.keepgoing.keepserver.domain.user.payload.request.SignupRequest;
 
 public interface UserService {
