+++ src/main/java/com/keepgoing/keepserver/domain/user/service/user/UserServiceImpl.java
@@ -1,8 +1,8 @@
 package com.keepgoing.keepserver.domain.user.service.user;
 
 import com.keepgoing.keepserver.domain.user.entity.user.User;
-import com.keepgoing.keepserver.domain.user.exception.BusinessException;
-import com.keepgoing.keepserver.domain.user.exception.error.ErrorCode;
+import com.keepgoing.keepserver.global.exception.BusinessException;
+import com.keepgoing.keepserver.global.exception.error.ErrorCode;
 import com.keepgoing.keepserver.domain.user.payload.request.SignupRequest;
 import com.keepgoing.keepserver.domain.user.payload.request.UserInfoRequest;
 import com.keepgoing.keepserver.domain.user.payload.response.JwtResponse;
